Stance Rework
Rewriting stances so that first and third person animations are the same, hands stay in the same position so don't clip into grips etc., and stuttering is eliminated
